refactor(cache-wrapper): move cache-wrapper export to use-cache package
- Remove cache-wrapper re-export from rari package to consolidate in use-cache
- Delete rari/src/runtime/cache-wrapper.ts and its export in vite.config.ts
- Move @rari/use-cache from dependencies to optionalDependencies in rari package
- Update all imports to reference @rari/use-cache/runtime/cache-wrapper directly
- Export deterministicStringify from use-cache/runtime/cache-wrapper.ts
- Update import path in transform.ts to use correct cache-wrapper location
- Add @rari/use-cache to root package.json devDependencies
- Update test imports to use new package path
- Add vite alias for @rari/use-cache/runtime/cache-wrapper in root config
- Consolidates cache utilities in single package for clearer dependency structure
