--- packages/rari/src/runtime/cache-wrapper.ts
@@ -1,2 +1,0 @@
-export { $$cache__, deterministicStringify, encodeBoundArgs } from '@rari/use-cache'
-export type { CacheEntry } from '@rari/use-cache'
