+++ packages/rari/vite.config.ts
@@ -14,7 +14,6 @@ export default defineConfig({
       'mdx': 'src/mdx.ts',
       'headers': 'src/headers.ts',
       'runtime/actions': 'src/runtime/actions.ts',
-      'runtime/cache-wrapper': 'src/runtime/cache-wrapper.ts',
       'runtime/entry-client': 'src/runtime/entry-client.ts',
       'runtime/react-server-dom-shim': 'src/runtime/react-server-dom-shim.ts',
       'runtime/rsc-client-runtime': 'src/runtime/rsc-client-runtime.ts',
