+++ packages/use-cache/src/runtime/cache-wrapper.ts
@@ -72,3 +72,5 @@ export function encodeBoundArgs(
 ): string {
   return serialize([refId, ...args]).toString('base64')
 }
+
+export { deterministicStringify }
