+++ vite.config.ts
@@ -4,6 +4,9 @@ import { defineConfig } from 'vite-plus'
 export default defineConfig({
   resolve: {
     alias: {
+      '@rari/use-cache/runtime/cache-wrapper': fileURLToPath(
+        new URL('./packages/use-cache/dist/runtime/cache-wrapper.mjs', import.meta.url),
+      ),
       '@rari/use-cache': fileURLToPath(
         new URL('./packages/use-cache', import.meta.url),
       ),
